+++ src/hooks/useTreasuryMetrics.ts
@@ -3,11 +3,11 @@ import { useTokenSuppliesQuery } from "src/generated/graphql";
 import { useCurrentIndex, useOhmPrice } from "src/hooks/useProtocolMetrics";
 import { useTokenRecordsLatestDate, useTreasuryLiquidValue } from "src/hooks/useTokenRecords";
 import { DEFAULT_RECORD_COUNT } from "src/views/TreasuryDashboard/components/Graph/Constants";
-import { getOhmCirculatingSupply } from "src/views/TreasuryDashboard/components/Graph/TokenSupplyQueryHelper";
+import { getOhmCirculatingSupply } from "src/views/TreasuryDashboard/components/Graph/helpers/TokenSupplyQueryHelper";
 import {
   getLiquidBackingPerGOhmSynthetic,
   getLiquidBackingPerOhmFloating,
-} from "src/views/TreasuryDashboard/components/Graph/TreasuryQueryHelper";
+} from "src/views/TreasuryDashboard/components/Graph/helpers/TreasuryQueryHelper";
 
 const QUERY_OPTIONS = { refetchInterval: 60000 }; // Refresh every 60 seconds
